% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{acf_length_func}
\alias{acf_length_func}
\title{acf_length_func}
\usage{
acf_length_func(fisher_df)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}
}
\value{
a single row tibble with the length of lag where acf is < e^-1
}
\description{
Calculates the autocorrelation function where its value is smaller than  e^−1
}
