% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{beyond_std_func}
\alias{beyond_std_func}
\title{beyond_std_func}
\usage{
beyond_std_func(fisher_df, sds = 1)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}

\item{sds}{Number of standard deviations to look beyond}
}
\value{
a single row tibble with percentage of points greater than n standard deviations
}
\description{
Calculates the percentage of poits beyone n standard deviations from the mean
}
