% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologic_login.R
\name{biologic_login}
\alias{biologic_login}
\title{biologic_login}
\usage{
biologic_login(email, password)
}
\arguments{
\item{email}{User's email address attached to Bioilogic account.}

\item{password}{Password for user's Biologic account.}
}
\description{
Login into Biomark's Biologic database to retreive an API token.
}
\examples{
biologic_login('email', 'password')
}
\author{
Ryan N. Kinzer
}
