% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{changepoint_func}
\alias{changepoint_func}
\title{changepoint_func}
\usage{
changepoint_func(fisher_df)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}
}
\value{
a single row tibble of changepoint summary
}
\description{
Uses the changepoint and changepoint.np packages to estimate changepoints using the meanvar and np method (see: http://members.cbio.mines-paristech.fr/~thocking/change-tutorial/RK-CptWorkshop.html)
}
