% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_func.R
\name{conversion_func}
\alias{conversion_func}
\title{Conversion function}
\usage{
conversion_func(df)
}
\arguments{
\item{df}{A dataframe containing time information and variables of interest}
}
\value{
A dataframe with each column as numeric data
}
\description{
Conversion function attempts to force input data structure to correct format (i.e., numeric data). *NOTE, does not attempt to rearrange columns or format dates*. Outputs must still be manually checked.
}
