% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisher}
\alias{fisher}
\title{Fisher function}
\usage{
fisher(
  df,
  sos = c(),
  w_size = 8,
  w_incre = 1,
  smooth_step = 3,
  RedRum = FALSE,
  write_out_csv = FALSE,
  write_out_rds = FALSE,
  display_plot = FALSE
)
}
\arguments{
\item{df}{A dataframe containing time information and variables of interest}

\item{sos}{An option vector containing size of state}

\item{w_size}{Window size}

\item{smooth_step}{Window size for the smoothing operation}

\item{w_incr}{Window increment}

\item{xtick_step}{Number of ticks on the x axis (currently unused)}
}
\value{
A dataframe where the last three columns are the Fisher's Information means, Fisher's Information smoothed and time-steps
}
\description{
The fisher function calculates the Fisher's Information following:
Cabezas, H., Campbell, D., Eason, T., Garmestani, A. S., Heberling, M. T., Hopton, M. E., Templeton, J., White, D., Zanowick, M., & Sparks, R. T. (2010). San Luis Basin sustainability metrics project: A methodology for evaluating regional sustainability. USEPA. USA.
Eason, T., & Cabezas, H. (2012). Evaluating the sustainability of a regional system using Fisher information in the San Luis Basin, Colorado. Journal of Environmental Management, 94(1), 41–49. https://doi.org/10.1016/j.jenvman.2011.08.003
Ahmad, N., Derrible, S., Eason, T., & Cabezas, H. (2016). Using Fisher information to track stability in multivariate systems. Royal Society Open Science, 3(11), 160582. https://doi.org/10.1098/rsos.160582
}
