% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{fisher_metrics}
\alias{fisher_metrics}
\title{fisher_metrics}
\usage{
fisher_metrics(fisher_df)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}
}
\value{
a single row tibble with the max number of windows in median state and mean state
}
\description{
Calculates the max number of windows in median state and mean state
}
