% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{frac_func}
\alias{frac_func}
\title{frac_func}
\usage{
frac_func(fisher_df)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}
}
\value{
a single row tibble with the fractions of increasing and decreasin first differences and ratio
}
\description{
Calculates the fractions of increasing and decreasin first differences and ratio
}
