% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biologic_data.R
\name{get_biologic_data}
\alias{get_biologic_data}
\title{get_biologic_data}
\usage{
get_biologic_data(
  site = NULL,
  endpoint = c("tags", "reader", "enviro", "antenna"),
  begin_dt = NULL,
  end_dt = NULL
)
}
\arguments{
\item{site}{IPTDS three character site code.}

\item{endpoint}{The type of data you want returned.}

\item{begin_dt}{Start date for the data returned.}

\item{end_dt}{End date for the data returned.}
}
\value{
data.frame
}
\description{
Pass API token to Biomark's Biologic database and receive a site
data.
}
\examples{
biologic_login('email', 'password')
data <- get_biologic_data('KRS', 'tags', '2023-06-01', '2023-06-02')
}
\author{
Ryan N. Kinzer
}
