% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biologic_sites.R
\name{get_biologic_sites}
\alias{get_biologic_sites}
\title{get_biologic_sites}
\usage{
get_biologic_sites()
}
\value{
vector
}
\description{
Pass API token to Biomark's Biologic database and receive a
vector of sites in return that the user has permissions.
}
\examples{
biologic_login('email', 'password')
sites <- get_biologic_sites()
}
\author{
Ryan N. Kinzer
}
