% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FINSdata.R
\name{get_FINSdata}
\alias{get_FINSdata}
\title{Download data from LSRCP FINS database.}
\usage{
get_FINSdata(
  module = c("Trapping", "Release", "Spawning"),
  scope = c("NPT", "FINS Domain"),
  startDate,
  endDate
)
}
\arguments{
\item{module}{FINS database module to query; Trapping, Release, Spawning.}

\item{scope}{download NPT data only or the full FINS domain}

\item{startDate}{query start date formatted as month, day, year; '01/01/2019'}

\item{endDate}{query end date formatted as month, day, year; '01/01/2020'}
}
\value{
a data frame of all returned data from FINS query
}
\description{
\code{get_FINSdata} downloads FINS data from their API using an
  NPT specific api key. The function can access data from three FINS modules;
  'Trapping', 'Release' and 'Spawning'. The user can also specify if only NPT
  data is required or the full dataset from FINS.
}
