% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RMIS_data.R
\name{get_RMIS_data}
\alias{get_RMIS_data}
\title{get_RMIS_data}
\usage{
get_RMIS_data(
  api_key,
  agency = c("NPT", "ADFG", "CCT", "CDFO", "CDFW", "CDFWKT", "CRITFC", "IDFG", "NMFS",
    "NMFSNWR", "NWIFC", "ODFW", "QDNR", "QUIL", "RMPC", "STIL", "USFWS", "WDFW", "YAKA",
    "YTFP"),
  module = c("release", "location", "files", "recovery", "catchsample", "description"),
  page = 1,
  perpage = 1000
)
}
\arguments{
\item{api_key}{unique access key obtained from RMIS.}

\item{agency}{desired return agency}

\item{module}{desired module to get data from}

\item{page}{desired page to return (default: 1)}

\item{perpage}{Number of records returned 1-1000 (default: 1000)}
}
\value{
data frame containing RMIS data from selected module
}
\description{
\code{get_RMIS_data} downloads data from the RMIS API returning
data for the specified agency and module. Restricted to returning a maximum
of 1000 records. If 1000 records are returned, more data should exist on
subsequent pages (Recommendation: increment your page argument).
}
