% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_USACE_data.R
\name{get_USACE_data}
\alias{get_USACE_data}
\title{get_USACE_data}
\source{
\url{https://www.nwd-wc.usace.army.mil/dd/common/dataquery/www/}
}
\usage{
get_USACE_data(
  query = c("BON.Power.Total.1Hour.1Hour.CBT-RAW", "TDA.Power.Total.1Hour.1Hour.CBT-RAW",
    "JDA.Power.Total.1Hour.1Hour.CBT-RAW", "MCN.Power.Total.1Hour.1Hour.CBT-RAW",
    "IHR.Power.Total.1Hour.1Hour.CBT-RAW", "LMN.Power.Total.1Hour.1Hour.CBT-RAW",
    "LGS.Power.Total.1Hour.1Hour.CBT-RAW", "LWG.Power.Total.1Hour.1Hour.CBT-RAW"),
  startdate,
  enddate,
  timezone = c("PST", "PDT", "MST", "MDT", "GMT")
)
}
\arguments{
\item{query}{Desired Dataquery 2.0 query name. Visit https://www.nwd-wc.usace.army.mil/dd/common/dataquery/www/
to find additional queries than those provided.}

\item{startdate}{Beginning date of query range provided as 'MM/DD/YYYY'}

\item{enddate}{Final date of query range provided as 'MM/DD/YYYY'}

\item{timezone}{Desired Timezone (options: PST, PDT, MST, MDT, GMT)}
}
\description{
Download data from the U.S. Army Corps of Engineers Dataquery 2.0 tool
}
\examples{
get_USACE_data(query = 'LWG.Power.Total.1Hour.1Hour.CBT-RAW',
startdate = '01/01/2022', enddate = '10/17/2022', timezone = 'PST')
}
\author{
Tyler Stright
}
