% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{magnitude_change_func}
\alias{magnitude_change_func}
\title{magnitude_change_func}
\usage{
magnitude_change_func(fisher_df, lags = 10)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}

\item{lags}{the number of lags to calculate the differences between}
}
\value{
a single row tibble with the max increase and max decrease found accros all lags
}
\description{
Calculates the differences between two point with a lag of 1:lags
}
