% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{median_BRP_func}
\alias{median_BRP_func}
\title{median_brp_func}
\usage{
median_BRP_func(fisher_df)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}
}
\value{
a single row tibble with the median buffer range percentage
}
\description{
Calculates the fraction of points within 10 percent of the range???
}
