% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{metrics_function}
\alias{metrics_function}
\title{metrics_function}
\usage{
metrics_function(fisher_df, stds = 1, lags = 10)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}

\item{stds}{number of standard deviations for beyond_std_func function}

\item{lags}{number of lags for magnitude_change_func function}
}
\value{
a single row tibble containing difference between the third and the first quartiles
}
\description{
Wrapper for all the metrics to calculate each one and return a single row tibble
}
