% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{percent_diff_percentile}
\alias{percent_diff_percentile}
\title{percent_diff_percentile}
\usage{
percent_diff_percentile(fisher_df)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}
}
\value{
a single row tibble containing the percent difference percentile
}
\description{
Calculates the difference between the 95th and 5th percentile over the median
}
