% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{percentile_ratio_mid_20}
\alias{percentile_ratio_mid_20}
\title{percentile_ratio_mid_20}
\usage{
percentile_ratio_mid_20(fisher_df)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}
}
\value{
a single row tibble with the mid 20 ratio
}
\description{
Calculates the mid 20 percentile ratio: percentile40,60 / percentileF5,95
}
