% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{percentile_ratio_mid_35}
\alias{percentile_ratio_mid_35}
\title{percentile_ratio_mid_35}
\usage{
percentile_ratio_mid_35(fisher_df)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}
}
\value{
a single row tibble with the mid 35 ratio
}
\description{
Calculates the mid 35 percentile ratio: percentile32.5,67.5 / percentileF5,95
}
