% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{percentile_ratio_mid_50}
\alias{percentile_ratio_mid_50}
\title{percentil_ratio_mid_50}
\usage{
percentile_ratio_mid_50(fisher_df)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}
}
\value{
a single row tibble with the mid 50 ratio
}
\description{
Calculates the mid 50 percentile ratio: percentile25,75 / percentileF5,95
}
