% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{percentile_ratio_mid_65}
\alias{percentile_ratio_mid_65}
\title{percentil_ratio_mid_65}
\usage{
percentile_ratio_mid_65(fisher_df)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}
}
\value{
a single row tibble with the mid 65 ratio
}
\description{
Calculates the mid 65 percentile ratio: percentile17.5,82.5 / percentileF5,95
}
