% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{percentile_ratio_mid_80}
\alias{percentile_ratio_mid_80}
\title{percentil_ratio_mid_80}
\usage{
percentile_ratio_mid_80(fisher_df)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}
}
\value{
a single row tibble mid 80 ratio
}
\description{
Calculates the mid 80 percentile ratio: percentile10,90 / percentileF5,95
}
