% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{q31_func}
\alias{q31_func}
\title{Q3-1}
\usage{
q31_func(fisher_df)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}
}
\value{
a single row tibble containing difference between the third and the first quartiles
}
\description{
Calculates the difference  between the third quartile, Q3, and the first quartile, Q1
}
