% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryPITtagAdult.R
\name{queryPITtagAdult}
\alias{queryPITtagAdult}
\title{PIT tag Adult Return Data}
\source{
\url{http://www.cbr.washington.edu/dart}
}
\usage{
queryPITtagAdult(
  site = c("GRA", "B2A"),
  species = c("Chinook", "Coho", "Steelhead", "Sockeye"),
  run = c("All", "Spring", "Summer", "Fall", "Winter", "Unknown"),
  rear_type = c("W+H", "Wild", "Hatchery", "Unknown"),
  spawn_yr = NULL,
  start_day = NULL,
  end_day = NULL
)
}
\arguments{
\item{spawn_yr}{Spawn year to query. Available years include 1988-current for GRA and 1995-current for B2A.}

\item{start_day}{Start date (\code{month / day}) of query.}

\item{end_day}{End date (\code{month / day}) of query.}

\item{dam}{The dam code to query for PIT tag data. Currently only available for Lower Granite Dam (GRA) and Bonneville (B2A).}

\item{spp}{Species to query. for window counts. Possible choices are: Chinook, Coho, Steelhead, Sockeye}
}
\description{
Query DART's PIT tag adult return observation detail.
}
\examples{

queryPITtagAdult(spawn_yr = 2015)
}
\author{
Ryan N. Kinzer and Kevin See
}
