% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryPITtagData.R
\name{queryPITtagData}
\alias{queryPITtagData}
\title{PIT tag data}
\source{
\url{http://www.cbr.washington.edu/dart}
}
\usage{
queryPITtagData(
  dam = "GRA",
  spp = c("Chinook", "Coho", "Steelhead", "Sockeye"),
  spawn_yr = NULL,
  start_day = NULL,
  end_day = NULL
)
}
\arguments{
\item{dam}{the dam code for the dam you wish to query for PIT tag data. Currently only available for Lower Granite Dam (GRA).}

\item{spp}{species to query window counts for. Possible choices are: Chinook, Coho, Steelhead, Sockeye}

\item{spawn_yr}{spawn year to query for window counts.}

\item{start_day}{date (\code{month / day}) when query should start}

\item{end_day}{date (\code{month / day}) when query should end}
}
\description{
Query and download PIT tag data about night passage and re-ascension at particular dams for specific species and year, using DART.
}
\examples{
queryPITtagData(spawn_yr = 2015)
}
\author{
Kevin See
}
