% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryRiverData.R
\name{queryRiverData}
\alias{queryRiverData}
\title{Columbia River Daily Environmental Datea}
\source{
\url{http://www.cbr.washington.edu/dart}
}
\usage{
queryRiverData(
  site = c("LWG", "LGS", "LMN", "IHR", "MCN", "JDA", "TDA", "BON"),
  year = NULL,
  start_day = NULL,
  end_day = NULL
)
}
\arguments{
\item{site}{The site code to query for PIT tag data. Currently only available for Lower Granite Dam (LWG) and Bonneville (BON) forebays.}

\item{year}{Year to query. Available years include 1979-current for LWG and 1949-current for BON.}

\item{start_day}{Start date (\code{month / day}) of query.}

\item{end_day}{End date (\code{month / day}) of query.}
}
\description{
Query DART's river environment daily data.
}
\examples{

queryPITtagAdult(spawn_yr = 2015)
}
\author{
Ryan N. Kinzer and Kevin See
}
