% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{rcs_func}
\alias{rcs_func}
\title{rcs_func}
\usage{
rcs_func(fisher_df)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}
}
\value{
a single row tibble with the range of the cumulative sum
}
\description{
Calculates the range of the cumulative sum divided by number of observations multiplied by the standard deviation
}
