% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{slope_rates_func}
\alias{slope_rates_func}
\title{summary_stats_func}
\usage{
slope_rates_func(fisher_df)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}
}
\value{
a single row tibble summarising slopes and rates
}
\description{
Calculates percent rate of change and slope between two consecutive points
}
