% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{small_kurtosis_func}
\alias{small_kurtosis_func}
\title{small_kurtosis_func}
\usage{
small_kurtosis_func(fisher_df)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}
}
\value{
a single row tibble with the small kurtosis
}
\description{
Calculates the kurtosis for small sample sizes
}
