% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sost.R
\name{sost}
\alias{sost}
\title{Size of state}
\usage{
sost(df, w_size = 8)
}
\arguments{
\item{df}{A dataframe containing time information and variables of interest}

\item{w_size}{Window size}
}
\value{
A vector containing size of state values with length equal to the number of variable columns in your input df
}
\description{
The sost function calculates the size of state following:
Cabezas, H., Campbell, D., Eason, T., Garmestani, A. S., Heberling, M. T., Hopton, M. E., Templeton, J., White, D., Zanowick, M., & Sparks, R. T. (2010). San Luis Basin sustainability metrics project: A methodology for evaluating regional sustainability. USEPA. USA.
Eason, T., & Cabezas, H. (2012). Evaluating the sustainability of a regional system using Fisher information in the San Luis Basin, Colorado. Journal of Environmental Management, 94(1), 41–49. https://doi.org/10.1016/j.jenvman.2011.08.003
Ahmad, N., Derrible, S., Eason, T., & Cabezas, H. (2016). Using Fisher information to track stability in multivariate systems. Royal Society Open Science, 3(11), 160582. https://doi.org/10.1098/rsos.160582
}
