% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metrics.R
\name{summary_stats_func}
\alias{summary_stats_func}
\title{Metric functions}
\usage{
summary_stats_func(fisher_df)
}
\arguments{
\item{fisher_df}{A fisher output containing FI_means and time_windows}
}
\value{
a single row tibble of summary statistics
}
\description{
These metric functions, currently in development, are made to collapse the Fisher's Information output into single number metrics
}
\details{
summary_stats_func

Calculates a range of masic summary statistics
}
