% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSeleniumTools.core.R
\name{web_automation_platform}
\alias{web_automation_platform}
\title{web automation platform}
\usage{
web_automation_platform(
  client_config = rst_remotedriver(get_config_list_only = TRUE)
)
}
\arguments{
\item{client_config}{config}
}
\description{
web automation platform
}
